Refactor Job.FileStore into fileStore.py (resolves #1028)
resolves #1028
