+++ src/toil/test/src/jobCacheTest.py
@@ -27,7 +27,8 @@ from abc import abstractmethod, ABCMeta
 from struct import pack, unpack
 from uuid import uuid4
 
-from toil.job import Job, IllegalDeletionCacheError
+from toil.job import Job
+from toil.fileStore import IllegalDeletionCacheError
 from toil.test import ToilTest, needs_aws, needs_azure, needs_google, experimental
 from toil.leader import FailedJobsException
 from toil.jobStores.abstractJobStore import NoSuchFileException
